pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turbine-py-client"
version = "0.1.0"
description = "Python client for Turbine CLOB prediction markets API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Turbine Markets", email = "engineering@turbine.markets"}
]
keywords = ["turbine", "prediction-markets", "clob", "orderbook", "ethereum", "trading"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Investment",
]
requires-python = ">=3.9"
dependencies = [
    "eth-account>=0.13.0",
    "eth-utils>=4.1.1",
    "httpx[http2]>=0.27.0",
    "websockets>=12.0",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "respx>=0.20.0",
]

[project.urls]
Homepage = "https://turbine.markets"
Documentation = "https://docs.turbine.markets"
Repository = "https://github.com/turbine/turbine-py-client"
Issues = "https://github.com/turbine/turbine-py-client/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["turbine_client*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
strict = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["turbine_client"]
